Allow player and radio group components to be aware of whether a locale is RTL
Precursor to reordering the radio button labels

[ci skip]
[#LEI-348]
